Move myTrade react-query hooks into src/queries/myTrades.ts
