+++ src/pages/Trade/components/OrderBookCard/AllTrades.tsx
@@ -1,33 +1,17 @@
 import React, { FC } from 'react';
-import { useQuery } from 'react-query';
-import snxData from 'synthetix-data';
 
-import { HistoricalTrades } from 'ducks/trades/types';
-
+import { useAllTradesQuery } from 'queries/myTrades';
 import TradeHistory from './TradeHistory';
-
-import { REFRESH_INTERVAL } from './constants';
-import QUERY_KEYS from 'constants/queryKeys';
 
 type AllTradesProps = {};
 
 const AllTrades: FC<AllTradesProps> = () => {
-	const allTradesQuery = useQuery<HistoricalTrades, any>(
-		QUERY_KEYS.HistoricalTrades.AllTrades,
-		() =>
-			snxData.exchanges.since({
-				maxBlock: Number.MAX_SAFE_INTEGER,
-				max: 100,
-			}),
-		{
-			refetchInterval: REFRESH_INTERVAL,
-		}
-	);
+	const allTradesQuery = useAllTradesQuery();
 
 	return (
 		<TradeHistory
 			trades={allTradesQuery.data || []}
-			isLoading={allTradesQuery.isLoading}
+			isLoading={allTradesQuery.isLoading && !allTradesQuery.isSuccess}
 			isLoaded={allTradesQuery.isSuccess}
 		/>
 	);
